Sync content of both "Contributing" pages
This updates the links in `specification/contributing.mdx` to point to
the new repo name, and changes `development/contributing.mdx` to mirror
`specification/contributing.mdx`.  Both files are kept because if the
"User Guide" tab and the "Specification" tab both link to the same file,
then Mintlify gets confused and always switches to the first tab when
navigating to that file.