Nest SDK links in "Get Started" group
Prior to this commit, the SDK links took up a lot of valuable real
estate at the top of the navigation sidebar.  Rather than being a
deliberate choice, their placement seemed to be an artifact of
Mintlify's limitations.  In Mintlify's configuration schema, external
links cannot be put into sidebar groups.  Instead they must be put in
the `anchors` section, which is always rendered at the top of the
sidebar.  Conversely, each entry in a group must refer to a page.

However, if a page's front matter contains the `url` property, Mintlify
will render links to that page as external links to the specified URL.
So this commit creates a stub page with a `url` property for each SDK,
and puts corresponding entries in the "Get Started" group, which are
then rendered as external links.