Move "SDKs" group to the bottom of the sidebar
This addresses a bug in the live production site wherein Mintlify will
try to navigate to an SDK's external URL when the user clicks on the
"SDKs" group in order to expand it.

This bug only occurs in the live production site, which is why it was
not caught during local development.