Use `<CodeGroup>` instead of `<Tabs>`
Mintlify assigns `id`s to tabs based solely on tab name.  Thus, if there
are multiple tabs with the same name on a page, clicking one of those
tabs will jump to the tab group containing the first such tab.  For
example, if there are two tab groups, each with a "TypeScript" tab and a
"Python" tab, clicking the "Python" tab in the second group will jump
to the first group where the "TypeScript" tab is still active.

This commit replaces `<Tabs>` with `<CodeGroup>` where applicable,
avoiding the problem and presenting a more appropriate UI.  The few
remaining occurrences of `<Tabs>` do not exihibit such conflicts, except
in the `docs/sdk/java/*.mdx` guides, which are currently not public.

Closes #1031.